More upgrades, split classes out
