+++ composer.json
@@ -10,7 +10,7 @@
         "php": ">=7.4.29"
     },
     "require-dev": {
-        "squizlabs/php_codesniffer": "^2.3",
+        "squizlabs/php_codesniffer": "^3.6.2",
         "php-parallel-lint/php-parallel-lint": "^1.3",
         "phpunit/phpunit": "^9",
         "php-parallel-lint/php-console-highlighter": "^1.0"
