+++ examples/left.php
@@ -11,10 +11,10 @@ require_once __DIR__ . '/../vendor/autoload.php';
 
 # minus($minuend, $subtrahend) is a left-associative operator.
 # e.g. "5 - 4 - 3" means "(5 - 4) - 3 = -2", not "5 - (4 - 3) = 4".
-function minus($minuend, $subtrahend)
-{
-    return $minuend - $subtrahend;
-}
+# function minus($minuend, $subtrahend)
+# {
+#     return $minuend - $subtrahend;
+# }
 
 # N -> number
 $N = new RegexParser(
